+++ R/api.R
@@ -1,5 +1,5 @@
 
-#' Query, set, delete list keys in a keyring
+#' Operations on keys
 #'
 #' These functions manipulate keys in a keyring. You can think of a keyring
 #' as a secure key-value store.
@@ -26,6 +26,20 @@
 #'
 #' `key_list` lists all keys of a keyring, or the keys for a certain
 #' service (if `service` is not `NULL`).
+#'
+#' ## Encodings
+#'
+#' On Windows, if required, an encoding can be specified using either
+#' an R option (`keyring.encoding_windows`) or environment variable
+#' (`KEYRING_ENCODING_WINDOWS`). This will be applied when both
+#' getting and setting keys. The option takes precedence over the
+#' environment variable, if both are set.
+#'
+#' This is reserved primarily for compatibility with keys set with
+#' other software, such as Python's implementation of keyring. For a
+#' list of encodings, use [iconvlist()], although it should be noted
+#' that not _every_ encoding can be properly converted, even for
+#' trivial cases. For best results, use UTF-8 if you can.
 #'
 #' @param service Service name, a character scalar.
 #' @param username Username, a character scalar, or `NULL` if the key
@@ -59,6 +73,28 @@
 #' key_get("R-keyring-test-service", "donaldduck")
 #' if (has_keyring_support()) key_list(service = "R-keyring-test-service")
 #' key_delete("R-keyring-test-service", "donaldduck")
+#'
+#' ## This is interactive using backend_file
+#' ## Set variables to be used in keyring
+#' kr_name <- "my_keyring"
+#' kr_service <- "my_database"
+#' kr_username <- "my_username"
+#'
+#' ## Create a keyring and add an entry using the variables above
+#' kb <- keyring::backend_file$new()
+#' ## Prompt for the keyring password, used to unlock keyring
+#' kb$keyring_create(kr_name)
+#' ## Prompt for the secret/password to be stored in the keyring
+#' kb$set(kr_service, username=kr_username, keyring=kr_name)
+#' # Lock the keyring
+#' kb$keyring_lock(kr_name)
+#'
+#' ## The keyring file is stored at ~/.config/r-keyring/ on Linux
+#'
+#' ## Output the stored password
+#' keyring::backend_file$new()$get(service = kr_service,
+#'   user = kr_username,
+#'   keyring = kr_name)
 #' }
 
 key_get <- function(service, username = NULL, keyring = NULL) {
@@ -124,7 +160,7 @@ key_list <- function(service = NULL, keyring = NULL) {
   default_backend()$list(service, keyring = keyring)
 }
 
-#' Manage keyrings
+#' Operations on keyrings
 #'
 #' On most platforms `keyring` supports multiple keyrings. This includes
 #' Windows, macOS and Linux (Secret Service) as well. A keyring is a