+++ R/backend-secret-service.R
@@ -1,5 +1,5 @@
 
-#' Create a Secret Service keyring backend
+#' Linux Secret Service keyring backend
 #'
 #' This backend is the default on Linux. It uses the libsecret library,
 #' and needs a secret service daemon running (e.g. Gnome Keyring, or
@@ -26,7 +26,7 @@
 #' \dontrun{
 #' ## This only works on Linux, typically desktop Linux
 #' kb <- backend_secret_service$new()
-#' kb$create_keyring("foobar")
+#' kb$keyring_create("foobar")
 #' kb$set_default_keyring("foobar")
 #' kb$set_with_value("service", password = "secret")
 #' kb$get("service")
@@ -45,7 +45,7 @@ backend_secret_service <- R6Class(
     get = function(service, username = NULL, keyring = NULL)
       b_ss_get(self, private, service, username, keyring),
     get_raw = function(service, username = NULL, keyring = NULL)
-      b_ss_get_raw(self, private, service, username = NULL, keyring = NULL),
+      b_ss_get_raw(self, private, service, username, keyring),
     set = function(service, username = NULL, keyring = NULL)
       b_ss_set(self, private, service, username, keyring),
     set_with_value = function(service, username = NULL, password = NULL,
@@ -101,15 +101,22 @@ b_ss_init <- function(self, private, keyring) {
 }
 
 b_ss_get <- function(self, private, service, username, keyring) {
-  keyring <- keyring %||% private$keyring
-  res <- .Call("keyring_secret_service_get", keyring, service, username)
+  res <- b_ss_get_raw(self, private, service, username, keyring)
   if (any(res == 0)) {
     stop("Key contains embedded null bytes, use get_raw()")
   }
   rawToChar(res)
 }
 
+b_ss_get_raw <- function(self, private, service, username, keyring) {
+  username <- username %||% getOption("keyring_username")
+  keyring <- keyring %||% private$keyring
+  res <- .Call(keyring_secret_service_get, keyring, service, username)
+  res
+}
+
 b_ss_set <- function(self, private, service, username, keyring) {
+  username <- username %||% getOption("keyring_username")
   password <- get_pass()
   b_ss_set_with_value(self, private, service, username, password, keyring)
   invisible(self)
@@ -117,29 +124,32 @@ b_ss_set <- function(self, private, service, username, keyring) {
 
 b_ss_set_with_value <- function(self, private, service, username, password,
                                 keyring) {
+  username <- username %||% getOption("keyring_username")
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_set", keyring, service, username,
+  .Call(keyring_secret_service_set, keyring, service, username,
         charToRaw(password))
   invisible(self)
 }
 
 b_ss_set_with_raw_value <- function(self, private, service, username, password,
                                     keyring) {
+  username <- username %||% getOption("keyring_username")
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_set", keyring, service, username,
+  .Call(keyring_secret_service_set, keyring, service, username,
         password)
   invisible(self)
 }
 
 b_ss_delete <- function(self, private, service, username, keyring) {
+  username <- username %||% getOption("keyring_username")
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_delete", keyring, service, username)
+  .Call(keyring_secret_service_delete, keyring, service, username)
   invisible(self)
 }
 
 b_ss_list <- function(self, private, service, keyring) {
   keyring <- keyring %||% private$keyring
-  res <- .Call("keyring_secret_service_list", keyring, service)
+  res <- .Call(keyring_secret_service_list, keyring, service)
   data.frame(
     service = res[[1]],
     username = res[[2]],
@@ -154,7 +164,7 @@ b_ss_keyring_create <- function(self, private, keyring) {
 }
 
 b_ss_keyring_list <- function(self, private) {
-  res <- .Call("keyring_secret_service_list_keyring")
+  res <- .Call(keyring_secret_service_list_keyring)
   data.frame(
     keyring = res[[1]],
     num_secrets = res[[2]],
@@ -166,26 +176,26 @@ b_ss_keyring_list <- function(self, private) {
 b_ss_keyring_delete <- function(self, private, keyring) {
   self$confirm_delete_keyring(keyring)
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_delete_keyring", keyring)
+  .Call(keyring_secret_service_delete_keyring, keyring)
   invisible()
 }
 
 b_ss_keyring_lock <- function(self, private, keyring) {
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_lock_keyring", keyring)
+  .Call(keyring_secret_service_lock_keyring, keyring)
   invisible()
 }
 
 b_ss_keyring_unlock <- function(self, private, keyring, password) {
   keyring <- keyring %||% private$keyring
   if (! is.null(password)) warning("password ignored, will be read interactively")
-  .Call("keyring_secret_service_unlock_keyring", keyring, password)
+  .Call(keyring_secret_service_unlock_keyring, keyring, password)
   invisible()
 }
 
 b_ss_keyring_is_locked <- function(self, private, keyring) {
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_is_locked_keyring", keyring)
+  .Call(keyring_secret_service_is_locked_keyring, keyring)
 }
 
 b_ss_keyring_default <- function(self, private) {
@@ -198,7 +208,7 @@ b_ss_keyring_set_default <- function(self, private, keyring) {
 }
 
 b_ss_is_available <- function(self, private, report_error) {
-  .Call("keyring_secret_service_is_available", report_error)
+  .Call(keyring_secret_service_is_available, report_error)
 }
 
 b_ss_keyring_create_direct <- function(self, private, keyring, password) {
@@ -206,6 +216,6 @@ b_ss_keyring_create_direct <- function(self, private, keyring, password) {
     warning("Password ignored, will be read interactively")
   }
   keyring <- keyring %||% private$keyring
-  .Call("keyring_secret_service_create_keyring", keyring)
+  .Call(keyring_secret_service_create_keyring, keyring)
   invisible(self)
 }